requirements.txt
fastapi[standard]==0.115.4
uvicorn[standard]>=0.22.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0
sqlalchemy>=2.0.0